refactor: rename configs
