--- src/config/main.ts
@@ -1,12 +1,0 @@
-import { ConfigService } from './config.service';
-
-let configSvc: ConfigService | undefined;
-
-export default function getConfig(): ConfigService {
-  if (configSvc) {
-    return configSvc;
-  }
-
-  configSvc = new ConfigService();
-  return configSvc;
-}
