+++ src/routes/browser/index.ts
@@ -4,14 +4,14 @@ import packageJSON from '../../package';
 import { engineList } from '../../handlers/main';
 import { indexSchema } from '../../types/requests/browser';
 
-import getConfig from '../../config/main';
+import env_config from '../../config/envConfig';
 
 export default async function indexRoute(fastify: FastifyInstance) {
   fastify.get('/', { schema: indexSchema }, async (_, reply) => {
     return reply.view('/templates/index.ejs', {
       packageJSON,
       engineList,
-      config: getConfig(),
+      config: env_config,
     });
   });
 }
